requirements.txt
langchain
langgraph
streamlit

langchain-community
ollama

langchain-nomic
nomic[local]

PyPDF2
fpdf

scikit-learn
joblib

typing-extensions
pypdf
python-dotenv